ALmost there, just touching up for accuracy and trying to understand this thoroughly, almost there lets go

+++ StudentGradeManager/Student.cs
@@ -2,8 +2,8 @@ namespace StudentGradeManager;
 
 public class Student
 {
-    public string Name { get; set; }
-    public int ID { get; set; }
+    public string Name { get; set; } 
+    public double ID { get; set; }
     public List<double> Grades { get; set; } =  new List<double>();
 
     public void AddGrade(double grade)
